backend/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
websockets==13.0
python-multipart==0.0.9
openai==1.50.0
PyMuPDF==1.24.10
python-docx==1.1.2
networkx==3.3
numpy==1.26.4
scikit-learn==1.5.2
httpx==0.27.2
aiofiles==24.1.0
